Bug#34574604 	ERROR 1712 (HY000): Index PRIMARY is corrupted
Background:
  When an online DDL is on-going and there are concurrent INSERTs,
UPDATEs or DELETEs on the table, the changes in the rows are logged in
the row log. After the DDL is complete, the row log is applied on the
new table. If an error occurs when applying the row logs, the DDL
reports the error and aborts.

  When applying the row log, if the base row contains any externally
stored BLOB columns, then any UPDATE in the row log is applied as a
DELETE followed by an INSERT. This is done first on the clustered index,
then on all secondary indexes. The entries to be DELETEd and INSERTed
are built from the row log entry. Any secondary index on virtual columns
will need the virtual column information in the row log to build the
correct entry, since the index entry contains the primary key value and
the computed virtual column value.

  The row log entry writes the virtual column information only when the
UPDATE changes the ordering field of the clustered index. An update not
changing the ordering field will not write virtual columns to row log.
This is irrespective of presence of secondary index on these columns.

Issue:
  Applying the row log fails in the scenario illustrated by the test
file. The failure occurs when the entry to be deleted in the secondary
index is not found, as it is not built correctly.

  The table contains a BLOB column, a virtually generated column and a
secondary index on the virtually generated column. The row being updated
has the BLOB stored externally. The test performs an online DDL by
adding a primary key, and concurrently updating the row such that the
ordering of the new primary key does not change.

  Since the update does not change the ordering field, the row log does
not contain the virtual column information. Since the row contains an
externally stored BLOB, the row log's UPDATE entry is applied as a
DELETE followed by an INSERT.

  The DELETE step fails when applied on the secondary index. The entry
to be deleted from the secondary index is built incorrectly, as the
virtual column value required is not present in the row log entry. Thus
the incorrect entry built is not found in the secondary index, causing
the DDL to abort.

Fix:
  Log the virutal column information if the row contains externally
stored BLOBs, either before or after the UPDATE.

Change-Id: I90bac31cf293565467b3f0b19741c8c5d646f63f