Bug#37065268 Null Pointer Exception in ReconnectTest
This patch for MySQL 8.4 and up reverts some changes that were
introduced in fixes for bug#29931625 "fix Java 11 deprecation
warnings" and bug#31710047 "distinct SessionFactories should not
share DomainTypeHandlers."

I begin by looking at SessionFactoryImpl in Cluster 7.4, where two
maps, proxyClassToDomainClass and typeToHandlerMap, are both marked
"TODO make this non-static." I understand this comment to be correct
in the second case (typeToHandlerMap), but incorrect in the first
one (proxyClassToDomainClass).

When Cluster/J creates an object for a user in session.newInstance(),
the object will belong to a proxy class. The proxyClassToDomainClass
map can associate the proxy class with some other class from the
user's code. This mapping is not tied to a specific database or
SessionFactory, and it does not become invalid after a disconnect.
It can safely be static, and the logic from bug#31710047 that cleared
this map after reconnect handling is unnecessary.

When the Java Module System was introduced in Java version 11,
Proxy.getProxyClass() was marked as deprecated (though not
for removal), because the returned class cannot be used in
other modules to create objects. This makes no difference at
all to Cluster/J, where the class is simply used as a hash key.
This patch reverts the work-around for getProxyClass() introduced
with bug#29931625.

All the synchronization needed to accees both maps is done in
SessionFactoryImpl, so this patch removes locking from
SessionImpl.

Change-Id: I66dce9b46e99f1ee3290c0d2f9f6b5921df4fbd8