PS-9434 database() results truncated at 34 bytes due to union
https://perconadev.atlassian.net/browse/PS-9434

Bug#36871927 mysql/mysql-server@2dd5e2ededb

The results are truncated because the metadata for function DATABASE()
indicates a string with maximum length 34 characters. This happens
because the length is taken from the constant value MAX_FIELD_NAME
instead of the correct NAME_CHAR_LEN (which gives the maximum length
of a typical identifier in the data dictionary).

The fix is to replace the wrong constant value.
In addition, the constant MAX_FIELD_NAME is removed since it is
basically unused. The only other use (in check_if_keyname_exists() to
synthesize a key name) is replaced with a value that makes the index
name use up to 64 single-byte characters, including underscore and
a generated two-digit sequence number.

Change-Id: I94b6327d0b8e867c9776afe3e68dd0e0ab83a939