PS-9434 FTS cleanup on DROP INDEX issue when adding a FULLTEXT index in same transaction
https://perconadev.atlassian.net/browse/PS-9434

Bug#36559642
mysql/mysql-server@5484c73368a
mysql/mysql-server@a8b3732507e
mysql/mysql-server@c25f8c03be8

Cause:
When all FULLTEXT indexes are dropped from a table, but another one
is added in the same transaction using INPLACE algorithm, the logic for
dropping the last FULLTEXT index performs cleanup on the assumption
that the in-memeory and on-disk FTS infrastructure is not needed any
more. Because of the way the INPLACE algorithm works, creation of said
infrastructure for the index being added and related checks will have
already been done at this point - the index is already added but not
yet published (the in-memory fts_t structure is not updated) and thus
not detected. This leaves the infrastructure in an inconsistent state,
leading to incorrect behavior.

Fix:
When dropping a FULLTEXT index, do not perform the cleanup if another
FULLTEXT index is to be added in the same transaction.

Test case added.

Change-Id: Iceed9f4c755d2cd44d2cf7b301fc0f3a18a521cc