Bug#37140331 MySQL NDB Cluster is crushing with Signal 8 error (Floating Point Exception) 1/2
Backport to 7.6

Problem:
When a query on ndbinfo table FRAG_MEM_USE or FRAG_OPERATIONS
is performed, a DBINFO_SCANREQ signal is sent to LQH. during
the handling of that signal, LQH asks TUP about fragStats
of all tables defined in the cluster.
If in 'parallel' with the ndbinfo scan a create (or drop)
table operation is on going, there could be a discrepancy
between the view LQH has of new table fragments and TUP view
about those fragments, that can lead to a crash in both TUP/ACC
or LQH.

In particular, if DBINFO_SCANREQ finds the new table with status
ADD_TABLE_ONGOING it could be impossible to TUP/ACC to get the
status of the new table fragments since at that point, fragments
information in TUP/ACC is not yet updated.
In similar way, during drop table if the status of the target table
in LQH is DROP_TABLE_* or PREP_DROP_* there could be differences
between the view that LQH and TUP/ACC have of the fragments of
that table.

This patch add 2 new tests to testNdbinfo suite where tables
FRAG_MEM_USE and FRAG_OPERATIONS are scanned (intensively) in
parallel with create/drop table operations running in different
threads.

Without a fix to the discrepancy in fragments in LQH and ACC/TUP
these tests should crash data nodes in different ways on both ACC
or TUP (it depends on the 'phase' of create/drop table at the
moment that SCANREQ of frags table arrives to LQH)

Note: due to some asserts in the code, if ndb(mt)d is compiled in
debug mode data node can crash earlier in those assert instead
of in the problematic places in LQH and ACC/TUP.

Change-Id: I40b2aed992d5086aeea711c162da88b5e1be19cb