BUG#37331118 Applier prepare error retry [test]
The replication applier normally retries temporary errors occurring
while applying transactions. Such retry logic is not performed for
transactions containing row events where the STMT_END_F flag is missing.
For such events, the statement will instead be committed as an
additional step while applying the subsequent COMMIT query event when it
is detected that there are still locked tables. When committing the
statement, temporary errors are not handled properly.

This patch reproduces the problem by writing an epoch trans with
simulated writes from multiple server ids on the source. The replica
then uses IGNORE_SERVER_IDS(<last_server_id_in_binlog>) to cause the
STMT_END_F to be filtered away, thus activating the above-described code
path in the applier. By holding a lock on one of the rows to be updated
by the applier, error handling is triggered. For reference the test then
also triggers error and retry handling when STMT_END_F has not been
filtered away.

Change-Id: Ifa3dc6d00691fbba286fcce5e1919f42a7e50d16