Bug#37162636 Ndb : GCP_SAVE_REF handling + recoverability [noclose]
Backport to 7.6

As described in the bug, there are some situations around graceful
shutdown + node failures which could lead to a cluster becoming
unrecoverable without manual intervention.

The causes of these situations can be addressed individually.

In this patch the generic GCI info propagation mechanism (CopyGCI)
is modified to reject propagating any set of GCI info which does
not describe the ability to automatically recover a cluster via
SR.

Specifically, it is essential that for a given restorable GCI,
there must be at least one node in every nodegroup also restorable
to that GCI.

Checking this in the Master before propagating it avoids it
becoming disk durable anywhere, so that in any cases where it
may be about-to-be violated, the problem does not propagate.

This could lead to a (faster) cluster shutdown, but the
cluster would then be automatically recoverable via SR.

Change-Id: I840267ef15873c47856ad31378052675efe008e8