BUG#37331118 Applier prepare error retry [compare_error]
The replication applier normally retries temporary errors occurring
while applying transactions. Such retry logic is not performed for
transactions containing row events where the STMT_END_F flag is missing.
For such events, the statement will instead be committed as an
additional step while applying the subsequent COMMIT query event when it
is detected that there are still locked tables. When committing the
statement, temporary errors are not handled properly.

This second patch fixes the handling of statement commit failure to
compare the returned error code in the same way as other failures. This
means that rather than attempting to commit the entire transaction (and
encountering an error indicating it has been rolled back), the applier
will now report the error and properly retry it as many times as
configured by the transaction retry settings.
In addition, the failure analysis capabilities are improved by extending
the error reporting to include all the warnings from the failed
statement commit, this is done in the same format and style as when
failing to apply row log events.

The effect of this change can be seen in the test implemented by the
prior patch:
 - the error code that stops replication is now the error from failing
   to commit the statement. Its also the same error code as when
   same error occurs without missing STMT_END_F.
 - the transaction retry attempt counter shows that the transaction was
   retried several times.

Each retry attempt output two lines to the MySQL Server log files, first
line indicates that error occured when committing statement and includes
all warnings, the second line only contains the actual error:
  <date> [Warning] [MY-010584] [Repl] Replica SQL for channel '':
    Worker 1 failed executing transaction 'ANONYMOUS' at source log
    binlog.000002, end_log_pos 711; Error in cleaning up after an event
    preceding the commit;  Got temporary error 266 'Time-out in NDB,
    probably caused by deadlock' from NDB, Error_code: 1297; Lock wait
    timeout exceeded; try restarting transaction, Error_code: 1205; Got
    error 146 - 'Lock timed out; Retry transaction' during COMMIT,
    Error_code: 1180; the group log file/position: binlog.000002
    0, Error_code: MY-000001
  <date> [Warning] [MY-010584] [Repl] Replica SQL for channel '':
    Worker 1 failed executing transaction 'ANONYMOUS' at source log
    binlog.000002, end_log_pos 711; Error 'Lock wait timeout exceeded;
    try restarting transaction' on query. Default database: ''. Query:
    'COMMIT', Error_code: MY-001205

Change-Id: I47337b8ae0841976219f70aa94d2861dd7b1ed6f