Bug#36234681: InnoDB: Failing assertion: result != FTS_INVALID
ANALYSIS

DELETE operation on a table with self referential foreign key
constraint and full text index may trigger an assert.

When performing the DELETE operation, marking of deletion of the
full text row happens twice due to the self referential foreign
key constraint. Hence the state of the row becomes invalid and
the assert is triggered.

Fix

Avoid updating the full text index if the foreign key has self
referential constraint.

NOTE: This is a 5.7 patch
Change-Id: I96d32937e8f50ecb580f3084c9fccd423408575d