Bug#36846567 Inplace ALTER TABLE might cause lost rows if concurrent purge [Post-push fix] Bug#37318367 Inplace ALTER on table with spatial idx might cause lost rows if concurrent purge [Post-push fix]
Description:
------------
This is post push fix of the above Bugs. There is another method
PCursor::move_to_next_block() where the scan is paused and latches are
released. This method was missed to be addressed, as a result
ALTER TABLE rebuild might lose records if latches are released in
this case.

Background:
-----------
When doing a parallel scan, Parallel_reader::Ctx::traverse() calls
Parallel_reader::Ctx::traverse_recs() which does:

```
  /* Note: The page end callback (above) can save and restore the cursor.
  The restore can end up in the middle of a page. */
  if (pcursor->is_after_last_on_page() && !move_to_next_node(pcursor)) {
    break;
  }
```
Where Parallel_reader::Ctx::move_to_next_node() calls
PCursor::move_to_next_block which uses PCursor::savepoint() and
PCursor::restore_from_savepoint().

These functions are known to have problems which were addressed in the
recent bug fixes, but they missed to fix this particular code path.
The reason for that omission was that the PCursor::move_to_next_block()
is also used during SELECT COUNT(*) at internal level (above leaves) for
which the solution employed for ALTER will not work, because B-tree
navigation functions do not support the required comparison mode when
searching to non-leaf level.

Fix:
----
This patch carves out the part of PCursor::move_to_next_block() which is
handling the simple leaf-level case, for which it is well known what
should be done: we are at supremum, all previous records were fully
processed, and we should restore the cursor to a user record which is
larger than any processed record. That can be accomplished by using the
recently introduced methods save_previous_user_record_as_last_processed()
and restore_to_first_unprocessed(), with a small adjustment, that here we
require the cursor to be not only after all processed records, but also
on a user record, so in case it is restored to supremum, we should
additionally call move_to_user_rec().

This patch also makes it clear which methods are used, and where, by
removing unused methods, marking some of them as private, and adding
asserts that m_read_level!=0 to those related to SELECT COUNT() only.

Change-Id: I2994374c542e168e0884e947634593ade17495ec